crs/codeql/requirements.txt
PyYAML>=6.0
Jinja2>=3.0
